Colorize_black_and_white_images/requirements.txt
colorize==1.1.0
numpy==2.2.1
opencv-python==4.10.0.84
pillow==11.0.0
